+++ packages/derived-wallet-base/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aptos-labs/derived-wallet-base
 
+## 0.7.0
+
+### Minor Changes
+
+- c066004: Bump wallet-standard and siwa packages
+
+## 0.6.0
+
+### Minor Changes
+
+- b6b9fc8: Fix DAA signing message and bump aptos ts-sdk version to 5.0.0
+
 ## 0.5.1
 
 ### Patch Changes